+++ src/time_split/__init__.py
@@ -11,6 +11,6 @@ __all__ = [
     "__version__",  # Make MyPy happy
 ]
 
-__version__ = "0.2.1.dev1"
+__version__ = "0.3.0"
 
 _logging.getLogger(__name__).addHandler(_logging.NullHandler())