+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-05-18
+
 ### Added
-- The `log_split_progress()`-function now accepts and optional _get_metrics_ callback method. Metrics returned are 
+- The `log_split_progress()`-function now accepts an optional _get_metrics()_ callback method. Metrics returned are 
   formatted based on type (if using the new `default_metrics_formatter()`), then added to the fold-end log message.
-- New property `settings.log_split_progress.FORMAT_METRICS`; formats _get_metrics_ output.
+- New property `settings.log_split_progress.FORMAT_METRICS`; formats _get_metrics()_ output.
 - Added several new `time_split.types` members, all related to `log_split_progress()`.
 
 ### Changed
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Fixed a few documentation and examples issues.
 
-[Unreleased]: https://github.com/rsundqvist/time-split/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/rsundqvist/time-split/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/rsundqvist/time-split/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/rsundqvist/time-split/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/rsundqvist/time-split/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/rsundqvist/time-split/compare/v0.0.0...v0.1.0